+++ applovin_max/lib/applovin_max.dart
@@ -13,7 +13,7 @@ export 'package:applovin_max/src/max_ad_view.dart';
 export 'package:applovin_max/src/max_native_ad_view.dart';
 
 /// The current version of the SDK.
-const String _version = "3.11.1";
+const String _version = "4.0.1";
 
 /// Represents the AppLovin SDK.
 class AppLovinMAX {
@@ -202,22 +202,6 @@ class AppLovinMAX {
     return _methodChannel.invokeMethod('hasUserConsent');
   }
 
-  /// Marks the user as age-restricted.
-  ///
-  /// [Prohibition on Personal Information from Children](https://developers.applovin.com/en/flutter/overview/privacy#prohibition-on-ads-to-and-personal-information-from-children-and-apps-exclusively-designed-for-or-exclusively-directed-to-children)
-  static void setIsAgeRestrictedUser(bool isAgeRestrictedUser) {
-    _methodChannel.invokeMethod('setIsAgeRestrictedUser', {
-      'value': isAgeRestrictedUser,
-    });
-  }
-
-  /// Checks if the user is age-restricted.
-  ///
-  /// [Prohibition on Personal Information from Children](https://developers.applovin.com/en/flutter/overview/privacy#prohibition-on-ads-to-and-personal-information-from-children-and-apps-exclusively-designed-for-or-exclusively-directed-to-children)
-  static Future<bool?> isAgeRestrictedUser() {
-    return _methodChannel.invokeMethod('isAgeRestrictedUser');
-  }
-
   /// Sets true to indicate that the user has opted out of interest-based advertising.
   ///
   /// Or, sets false to indicate that the user has not opted out of interest-based advertising.
@@ -368,7 +352,7 @@ class AppLovinMAX {
   //
 
   /// Sets an [AdViewAdListener] listener with which you can receive notifications about ad events.
-  static void setBannerListener(AdViewAdListener listener) {
+  static void setBannerListener(AdViewAdListener? listener) {
     _bannerAdListener = listener;
   }
 
@@ -484,7 +468,7 @@ class AppLovinMAX {
   //
 
   /// Sets an [AdViewAdListener] listener with which you can receive notifications about ad events.
-  static void setMRecListener(AdViewAdListener listener) {
+  static void setMRecListener(AdViewAdListener? listener) {
     _mrecAdListener = listener;
   }
 
@@ -574,7 +558,7 @@ class AppLovinMAX {
   //
 
   /// Sets an [InterstitialListener] listener with which you can receive notifications about ad events.
-  static void setInterstitialListener(InterstitialListener listener) {
+  static void setInterstitialListener(InterstitialListener? listener) {
     _interstitialListener = listener;
   }
 
@@ -619,7 +603,7 @@ class AppLovinMAX {
   //
 
   /// Sets a [RewardedAdListener] listener with which you can receive notifications about ad events.
-  static void setRewardedAdListener(RewardedAdListener listener) {
+  static void setRewardedAdListener(RewardedAdListener? listener) {
     _rewardedAdListener = listener;
   }
 
@@ -664,7 +648,7 @@ class AppLovinMAX {
   //
 
   /// Sets a [AppOpenAdListener] listener with which you can receive notifications about ad events.
-  static void setAppOpenAdListener(AppOpenAdListener listener) {
+  static void setAppOpenAdListener(AppOpenAdListener? listener) {
     _appOpenAdListener = listener;
   }
 
@@ -707,7 +691,7 @@ class AppLovinMAX {
   /// Sets a [WidgetAdViewAdListener] to receive notifications about
   /// [MaxAdView] ad events when preloading a [MaxAdView] platform widget with
   /// [preloadWidgetAdView].
-  static void setWidgetAdViewAdListener(WidgetAdViewAdListener listener) {
+  static void setWidgetAdViewAdListener(WidgetAdViewAdListener? listener) {
     _widgetAdViewAdListener = listener;
   }