+++ applovin_max/CHANGELOG.md
@@ -1,5 +1,13 @@
 ## Versions
 
+## 4.0.1
+* Crash if plugin version is incompatible with native SDK version.
+* Add support for Yandex native ads.
+* Update listener APIs to allow for null values for unsetting.
+## 4.0.0
+* Depends on Android SDK v13.0.0 and iOS SDK v13.0.0.
+* Removed COPPA support.
+* Add constants for MAX Error Codes. For more info, check out our [docs](https://developers.applovin.com/en/max/flutter/overview/error-handling#max-error-codes).
 ## 3.11.1
 * Depends on Android SDK v12.6.1 and iOS SDK v12.6.1.
 * Add APIs to preload widgets for banners and MRECs. For more info, check out our [docs](https://developers.applovin.com/en/max/flutter/ad-formats/banner-mrec-ads#ad-preloading).